requirements.txt
aiogram>=3.15.0
python-dotenv>=1.0.0
requests>=2.32.0
gspread==6.0.0
gspread-asyncio>=2.0.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pandas>=2.0.0
openpyxl>=3.1.0
apscheduler>=3.10.0  # ДОБАВЛЕНО: планировщик для авто-отчётов
